### 2025-05-29 #8306 @editorialbot
[PRE REVIEW]: pyMassEvac: A Python package for simulating multi-domain mass evacuation operations
**Submitting author:** @mark-rempel-gc (<a href="http://orcid.org/0000-0002-6248-1722">Mark Rempel</a>)
**Repository:** https://github.com/DRDC-RDDC/pyMassEvac
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
