+++ paper/paper.bib
@@ -108,7 +108,8 @@
   institution = {University Stravanger},
   year = {2016},
   number = {Report 58},
-  note = {https://uis.brage.unit.no/uis-xmlui/bitstream/handle/11250/2414815/Rapport_58.pdf?sequence=3&isAllowed=y}
+  note = {https://uis.brage.unit.no/uis-xmlui/bitstream/handle/11250/2414815/Rapport_58.pdf?sequence=3&isAllowed=y},
+  doi = {10.31265/usps.210}
 }
 
 @techreport{solberg2018a,
@@ -117,7 +118,8 @@
   institution = {University Stravanger},
   year = {2018},
   number = {Report 75},
-  note = {https://uis.brage.unit.no/uis-xmlui/handle/11250/2578301}
+  note = {https://uis.brage.unit.no/uis-xmlui/handle/11250/2578301}.
+  doi = {10.31265/usps.21}
 }
 
 @misc{mcnutt2016a,
