### 2025-05-29 38b41e0
Fix typo in .bib file
