+++ paper/paper.bib
@@ -118,7 +118,7 @@
   institution = {University Stravanger},
   year = {2018},
   number = {Report 75},
-  note = {https://uis.brage.unit.no/uis-xmlui/handle/11250/2578301}.
+  note = {https://uis.brage.unit.no/uis-xmlui/handle/11250/2578301},
   doi = {10.31265/usps.21}
 }
 
