### 2025-06-03 1999191
Update README.rst
Added badges
