### 2025-09-19 #8306 @arfon
@oak97 @leonlan @parikshitbajpai :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _pyMassEvac: A Python package for simulating multi-domain mass evacuation operations_

Many thanks
Arfon
