### 2025-09-19 #8306 @parikshitbajpai
Hi @arfon,

Thanks for the invitation to review this. I had a quick look at the paper and the code and this looks interesting. I can review it but I will only be able to do it starting the first week of October.

Cheers,
Parikshit
