### 2025-09-20 #9038 @arfon
Reminder to future self 😸  that @parikshitbajpai informed us they won't be able to start their review until early October.
