### 2025-12-15 1bf1860
Add test running instructions and minor typo fixes (refs. https://github.com/openjournals/joss-reviews/issues/9038)
