### 2025-12-15 #9038 @parikshitbajpai
Hi @arfon,

I am done with my review and I created a small pull request over in the project repository to fix a couple of typos I found and add instructions to run the tests. I think this is a well done software and the need for it has been clearly communicated by @mark-rempel-gc in the paper. The code is well documented and the tutorials provide a good place for new users to start using the package.

@mark-rempel-gc, please let me know if you have any questions about the PR and feel free to ignore any changes you don't agree with. Also, is there a plan to host the documentation website somewhere?
