### 2025-12-16 #9038 @mark-rempel-gc
@parikshitbajpai - I reviewed your suggested changes, accepted all of them, and merged them into the main branch. Thank you for your thorough review and positive feedback.

I expect that any documentation associated with this package will either be hosted on DRDC's GitHub site (https://github.com/DRDC-RDDC) through the package's site itself, or if written as a formal DRDC report will be published on our publications website (https://pubs.drdc-rddc.gc.ca/pubdocs/pcow1_e.html) - ideally both.
