### 2025-12-19 96af38d
Minor edits
