+++ paper/paper.bib
@@ -57,7 +57,7 @@
 @article{rempel2023a,
 	author = {M. Rempel and N. Shiell},
 	url = {https://review.sto.nato.int/},
-	journal = {NATO STO Review},
+	journal = {NATO Journal of Science and Technology},
 	pages = {21-1--21-20},
 	title = {Using Reinforcement Learning to Provide Decision Support in Multi-Domain Mass Evacuation Operations},
 	volume = {Fall 2023},
