### 2025-12-19 #9038 @mark-rempel-gc
> [@mark-rempel-gc](https://github.com/mark-rempel-gc) – looks like we're very close to being done here. I will circle back here next week, but in the meantime, please give your own paper a final read to check for any potential typos etc.

@arfon  - I have reviewed the paper, readme, and tutorial. I fixed a few minor typos, and they are ready for publication.

> After that, could you make a new release of this software that includes the changes that have resulted from this review.

I tried to make a release of the software, which seem to work fine. However, the draft-pdf action runs when I do so, and fails with the error message: "Error: fatal: You are not currently on a branch. To push the history leading to the current (detached HEAD)". I have not had this issue before and am at a bit of a loss as to how to fix it. That said, the PDF is correct in the repository itself as it compiled when I made the last set of changes to the source.

Any advice you could provide would be greatly appreciated.
