+++ .github/workflows/draft-pdf.yml
@@ -5,6 +5,9 @@ on:
       - paper/**
       - .github/workflows/draft-pdf.yml
 
+permissions:
+  contents: write
+
 jobs:
   paper:
     runs-on: ubuntu-latest
